pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pizerocam"
version = "0.1.0"
description='Python utils package for self-driving labs'
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0"
]

# More at https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 1 - Planning",
    "License :: MIT",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development",
]

authors = [
    {name = "Yang Cao", email = "cyrilcaoyang@gmail.com"}
]

[project.optional-dependencies]
server = [
    "picamera2",
    "neopixel",
    "RPi.GPIO"
]
client = [
    "opencv-python>=4.10.0",
    "numpy==2.2.2",
    "pillow==11.1.0",
    "matplotlib",
    "google-cloud-vision",
    "python-dotenv"
]
client-tesseract = [
    "opencv-python>=4.10.0",
    "numpy==2.2.2",
    "pillow==11.1.0",
    "pytesseract==0.3.13",
    "matplotlib",
    "python-dotenv"
]

[project.scripts]
pizerocam-client = "image_req_client.cli:main_client"
pizerocam-server = "image_server.cli:main_server"

[tool.setuptools]
packages = ["image_req_client", "image_server"]

[tool.setuptools.package-dir]
"" = "src"
